+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Error when communicating with `vtex.search-graphql` breaking all queries related to the order form.
 
 ## [0.36.1] - 2020-07-08
 ### Fixed